Rename to test_setup.rs
